+++ ui/components/multichain/pages/review-permissions-page/review-permissions-page.tsx
@@ -7,14 +7,7 @@ import {
   NonEmptyArray,
   parseCaipAccountId,
   KnownCaipNamespace,
-  Hex,
 } from '@metamask/utils';
-import type {
-  GatorPermissionsMap,
-  StoredGatorPermissionSanitized,
-  SignerParam,
-  PermissionTypes,
-} from '@metamask/gator-permissions-controller';
 import { uniq } from 'lodash';
 import { CAIP_FORMATTED_EVM_TEST_CHAINS } from '../../../../../shared/constants/network';
 import { endTrace, trace, TraceName } from '../../../../../shared/lib/trace';
@@ -67,157 +60,9 @@ import {
   EvmAndMultichainNetworkConfigurationsWithCaipChainId,
   MergedInternalAccountWithCaipAccountId,
 } from '../../../../selectors/selectors.types';
-import { getGatorPermissionsMap } from '../../../../selectors/gator-permissions/gator-permissions';
+import { GatorPermissionState, getFilteredGatorPermissionsByType } from '../../../../selectors/gator-permissions/gator-permissions';
 import { SiteCell } from './site-cell/site-cell';
 import { PermissionsCell } from './permissions-cell/permissions-cell';
-
-/**
- * Filters and groups gator permissions by type and site origin, returning counts and chain lists.
- *
- * This function takes the full gator permissions map and filters it to return
- * only permissions that match the specified site origin.
- * It groups the results into two main categories:
- * - streams: Combined count of native-token-stream and erc20-token-stream permissions
- * - subscriptions: Combined count of native-token-periodic and erc20-token-periodic permissions
- * Each category includes a list of all chains on which these permissions occur.
- *
- * @param gatorPermissions - The gator permissions map to filter
- * @param siteOrigin - The site origin to filter by (e.g., 'https://example.com')
- * @returns Object with counts and chain lists for streams and subscriptions
- */
-const getFilteredGatorPermissionsByType = (
-  gatorPermissions: GatorPermissionsMap,
-  siteOrigin: string,
-) => {
-  const result = {
-    streams: {
-      count: 0,
-      chains: new Set<Hex>(),
-      permissions: [] as {
-        permission: StoredGatorPermissionSanitized<
-          SignerParam,
-          PermissionTypes
-        >;
-        chainId: Hex;
-        permissionType: string;
-      }[],
-    },
-    subscriptions: {
-      count: 0,
-      chains: new Set<Hex>(),
-      permissions: [] as {
-        permission: StoredGatorPermissionSanitized<
-          SignerParam,
-          PermissionTypes
-        >;
-        chainId: Hex;
-        permissionType: string;
-      }[],
-    },
-  };
-
-  // Process stream permissions (native-token-stream + erc20-token-stream)
-  const streamTypes: (keyof GatorPermissionsMap)[] = [
-    'native-token-stream',
-    'erc20-token-stream',
-  ];
-  streamTypes.forEach((permissionType) => {
-    const permissionsForType = gatorPermissions[permissionType];
-    if (permissionsForType) {
-      Object.entries(permissionsForType).forEach(([chainId, permissions]) => {
-        const chainIdHex = chainId as Hex;
-        // Filter permissions by site origin
-        const filteredPermissions = permissions.filter(
-          (
-            permission: StoredGatorPermissionSanitized<
-              SignerParam,
-              PermissionTypes
-            >,
-          ) => permission.siteOrigin.toLowerCase() === siteOrigin.toLowerCase(),
-        );
-
-        if (filteredPermissions.length > 0) {
-          result.streams.count += filteredPermissions.length;
-          result.streams.chains.add(chainIdHex);
-
-          // Add raw permission data only
-          filteredPermissions.forEach(
-            (
-              permission: StoredGatorPermissionSanitized<
-                SignerParam,
-                PermissionTypes
-              >,
-            ) => {
-              result.streams.permissions.push({
-                permission,
-                chainId: chainIdHex,
-                permissionType,
-              });
-            },
-          );
-        }
-      });
-    }
-  });
-
-  // Process subscription permissions (native-token-periodic + erc20-token-periodic)
-  const subscriptionTypes: (keyof GatorPermissionsMap)[] = [
-    'native-token-periodic',
-    'erc20-token-periodic',
-  ];
-  subscriptionTypes.forEach((permissionType) => {
-    const permissionsForType = gatorPermissions[permissionType];
-    if (permissionsForType) {
-      Object.entries(permissionsForType).forEach(([chainId, permissions]) => {
-        const chainIdHex = chainId as Hex;
-        // Filter permissions by site origin
-        const filteredPermissions = permissions.filter(
-          (
-            permission: StoredGatorPermissionSanitized<
-              SignerParam,
-              PermissionTypes
-            >,
-          ) => permission.siteOrigin.toLowerCase() === siteOrigin.toLowerCase(),
-        );
-
-        if (filteredPermissions.length > 0) {
-          result.subscriptions.count += filteredPermissions.length;
-          result.subscriptions.chains.add(chainIdHex);
-
-          // Add raw permission data only
-          filteredPermissions.forEach(
-            (
-              permission: StoredGatorPermissionSanitized<
-                SignerParam,
-                PermissionTypes
-              >,
-            ) => {
-              result.subscriptions.permissions.push({
-                permission,
-                chainId: chainIdHex,
-                permissionType,
-              });
-            },
-          );
-        }
-      });
-    }
-  });
-
-  // Convert Sets to arrays for the final result
-  return {
-    streams: {
-      count: result.streams.count,
-      chains: Array.from(result.streams.chains),
-      permissions: result.streams.permissions,
-    },
-    subscriptions: {
-      count: result.subscriptions.count,
-      chains: Array.from(result.subscriptions.chains),
-      permissions: result.subscriptions.permissions,
-    },
-  };
-};
 
 export const ReviewPermissions = () => {
   const t = useI18nContext();
@@ -315,12 +160,10 @@ export const ReviewPermissions = () => {
     getAllPermittedChainsForSelectedTab(state, activeTabOrigin),
   ) as CaipChainId[];
 
-  const gatorPermissions = useSelector(getGatorPermissionsMap);
-
   // Get filtered gator permissions grouped by type
-  const filteredGatorPermissions = useMemo(() => {
-    return getFilteredGatorPermissionsByType(gatorPermissions, activeTabOrigin);
-  }, [gatorPermissions, activeTabOrigin]);
+  const filteredGatorPermissions = useSelector((state: GatorPermissionState) =>
+    getFilteredGatorPermissionsByType(state, activeTabOrigin)
+  );
 
   // Get all gator permissions for the site (for the modal)
   const allSiteGatorPermissions = useMemo(() => {
